pyproject.toml
[project]
name = "laughing-man"
version = "0.1.0"
description = "Webcam Laughing Man face overlay (MediaPipe + OpenCV + Pillow)"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "mediapipe>=0.10.33",
    "pillow>=12.2.0",
    "pyvirtualcam>=0.15.0",
    "typer>=0.24.1",
]

[project.scripts]
laughing-man = "laughing_man:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/laughing_man"]

# uv defaults to hardlinks from its cache; when cache and .venv are on different
# devices, that can fail silently and leave broken wheels (e.g. cv2 without
# VideoCapture). Copying avoids incomplete OpenCV installs.
[tool.uv]
link-mode = "copy"
